Split CtrfReportMerger into focused partial files
Refactor the 777-line CtrfReportMerger.cs into four focused partial class
files, each under 300 lines, with no behavioral change:

- CtrfReportMerger.cs: merge orchestration (Merge overloads)
- CtrfReportMerger.FileMerge.cs: file-based entry points (MergeToFileAsync)
- CtrfReportMerger.RetryCollapsing.cs: CollapseRetryAttempts pipeline
- CtrfReportMerger.JsonHelpers.cs: JSON readers, environment merge, report id

Fixes #10347

Copilot-Session: 654b45b0-5318-4a77-8ef3-2434c5226c09
