+++ src/Platform/Microsoft.Testing.Extensions.CtrfReport/CtrfReportMerger.cs
@@ -4,8 +4,6 @@
 using System.Text.Json;
 using System.Text.Json.Nodes;
 
-using Microsoft.Testing.Platform;
-
 namespace Microsoft.Testing.Extensions.CtrfReport;
 
 /// <summary>
@@ -37,7 +35,7 @@ namespace Microsoft.Testing.Extensions.CtrfReport;
 /// validator or linter.
 /// </para>
 /// </remarks>
-internal static class CtrfReportMerger
+internal static partial class CtrfReportMerger
 {
     // Neutral tool identity used when merged inputs disagree on their producing test framework, so the
     // merged report does not misattribute one framework's identity to another's tests.
@@ -46,24 +44,6 @@ internal static class CtrfReportMerger
     // Identity stamped into the merged document's 'generatedBy' — the merge is produced by this extension,
     // not by any input report.
     private const string GeneratedByName = "Microsoft.Testing.Extensions.CtrfReport";
-
-    // Fields a CTRF retry attempt object (section 11) shares with a test object and that carry over verbatim when
-    // a non-final attempt is folded into 'retryAttempts[]'. 'attempt' and 'status' are handled separately because
-    // they are required, and 'attemptId' is listed here because an input that assigned one keeps it.
-    private static readonly string[] RetryAttemptFields =
-    [
-        "attemptId",
-        "duration",
-        "message",
-        "trace",
-        "line",
-        "snippet",
-        "stdout",
-        "stderr",
-        "start",
-        "stop",
-        "attachments",
-    ];
 
     internal static string Merge(IReadOnlyList<string> inputReports)
         => Merge(inputReports, CtrfMergeMode.Concatenate);
@@ -305,473 +285,4 @@ internal static class CtrfReportMerger
 
         return merged.ToJsonString(new JsonSerializerOptions { WriteIndented = true });
     }
-
-    internal static Task MergeToFileAsync(
-        IReadOnlyList<string> inputPaths,
-        string outputPath,
-        CancellationToken cancellationToken)
-        => MergeToFileAsync(inputPaths, outputPath, CtrfMergeMode.Concatenate, cancellationToken);
-
-    internal static async Task MergeToFileAsync(
-        IReadOnlyList<string> inputPaths,
-        string outputPath,
-        CtrfMergeMode mode,
-        CancellationToken cancellationToken)
-    {
-        if (inputPaths is null)
-        {
-            throw new ArgumentNullException(nameof(inputPaths));
-        }
-
-        if (outputPath is null)
-        {
-            throw new ArgumentNullException(nameof(outputPath));
-        }
-
-        // Reject an empty input list before any filesystem work (Merge throws for empty input, but only
-        // after the output directory would already have been created).
-        if (inputPaths.Count == 0)
-        {
-            throw new ArgumentException("At least one CTRF report is required to merge.", nameof(inputPaths));
-        }
-
-        // RFC 018 treats per-module inputs as read-only and requires them to remain on disk; reject an
-        // output that aliases an input so a merge can never overwrite one of its own sources.
-        MergeOutputFileHelper.EnsureOutputDoesNotAliasInput(inputPaths, outputPath);
-
-        var reports = new List<string>(inputPaths.Count);
-        foreach (string inputPath in inputPaths)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-#if NETCOREAPP
-            reports.Add(await File.ReadAllTextAsync(inputPath, cancellationToken).ConfigureAwait(false));
-#else
-            reports.Add(File.ReadAllText(inputPath));
-#endif
-        }
-
-        string merged = Merge(reports, mode);
-
-        string? outputDirectory = Path.GetDirectoryName(outputPath);
-        if (!RoslynString.IsNullOrEmpty(outputDirectory))
-        {
-            Directory.CreateDirectory(outputDirectory);
-        }
-
-        // Write to a temporary sibling, then replace the destination ENTRY, so a symlink/hardlink output
-        // alias of an input has only its link removed rather than the read-only source truncated in place.
-        await MergeOutputFileHelper.WriteViaTemporarySiblingAsync(outputPath, async tempPath =>
-        {
-#if NETCOREAPP
-            await File.WriteAllTextAsync(tempPath, merged, cancellationToken).ConfigureAwait(false);
-#else
-            File.WriteAllText(tempPath, merged);
-            await Task.CompletedTask.ConfigureAwait(false);
-#endif
-        }).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Folds successive attempts of the same logical test — the tests[] rows contributed, in attempt order, by
-    /// the per-attempt documents of one orchestrated retry run — into a single row per test.
-    /// </summary>
-    /// <remarks>
-    /// The retry model confirmed in ctrf-io/ctrf#58: the last attempt's outcome IS the test object, and
-    /// <c>retryAttempts[]</c> holds attempts <c>1..N-1</c> (the initial execution plus any earlier retries),
-    /// numbered from 1, so <c>retries == retryAttempts.length</c> and the final attempt is <c>retries + 1</c>.
-    /// The test object therefore keeps the FINAL attempt's <c>duration</c> (and <c>start</c>/<c>stop</c>), rather
-    /// than a sum across attempts. Attempts an input already recorded in its own <c>retryAttempts[]</c> (in-process
-    /// retries within a single attempt process) are flattened into the same history so no execution is lost.
-    /// </remarks>
-    private static JsonArray CollapseRetryAttempts(JsonArray tests)
-    {
-        var slots = new List<(JsonObject Final, List<JsonObject> Priors)>();
-        var byIdentity = new Dictionary<string, int>(StringComparer.Ordinal);
-
-        foreach (JsonNode? test in tests)
-        {
-            // Non-objects were already rejected during ingestion; this only re-establishes the type.
-            if (test is not JsonObject testObject)
-            {
-                continue;
-            }
-
-            // A row we cannot identify gets its own slot: fusing unrelated rows would lose results, whereas an
-            // uncollapsed duplicate is merely redundant.
-            if (GetTestIdentity(testObject) is not string identity)
-            {
-                slots.Add((testObject, []));
-                continue;
-            }
-
-            if (byIdentity.TryGetValue(identity, out int index))
-            {
-                (JsonObject previousFinal, List<JsonObject> priors) = slots[index];
-                priors.Add(previousFinal);
-                slots[index] = (testObject, priors);
-            }
-            else
-            {
-                byIdentity.Add(identity, slots.Count);
-                slots.Add((testObject, []));
-            }
-        }
-
-        var collapsed = new JsonArray();
-        foreach ((JsonObject final, List<JsonObject> priors) in slots)
-        {
-            collapsed.Add(BuildCollapsedTest(final, priors));
-        }
-
-        return collapsed;
-    }
-
-    /// <summary>
-    /// Computes the key identifying the logical test a row describes, preferring the CTRF <c>testId</c>, then the
-    /// legacy <c>id</c>, then the producer-supplied <c>extra.uid</c>, and finally the suite path plus name and the
-    /// other stable discriminators the Test object offers. Returns <see langword="null"/> when the row carries
-    /// none of them.
-    /// </summary>
-    /// <remarks>
-    /// <para>
-    /// The identifier order follows CTRF 9.1: <c>id</c> is a stable test-case identifier that consumers treat as
-    /// legacy, using <c>testId</c> in preference only when both are present. Ignoring an <c>id</c>-only report
-    /// would drop it to the heuristic fallback and risk fusing distinct same-named tests.
-    /// </para>
-    /// <para>
-    /// The fallback length-prefixes every component rather than just separating them. A CTRF <c>name</c> or suite
-    /// segment is an arbitrary non-empty string, so it may itself contain the separator: with plain separation,
-    /// <c>suite: ["A"], name: "B\u001fC"</c> and <c>suite: ["A", "B"], name: "C"</c> would produce the same key
-    /// and collapse two unrelated tests into one, silently dropping a result. Length prefixes make the encoding
-    /// unambiguous whatever the components contain.
-    /// </para>
-    /// <para>
-    /// The fallback also folds in <c>filePath</c> (9.19) and <c>parameters</c> (9.30), because suite plus name
-    /// alone is not unique: parameterized rows share both while differing only in their parameters, and
-    /// same-named tests in different files differ only by path. If a producer were to serialize <c>parameters</c>
-    /// inconsistently between attempts, the effect is that a retry is not folded — a duplicated row rather than a
-    /// lost result, which is the safe direction to fail in.
-    /// </para>
-    /// </remarks>
-    private static string? GetTestIdentity(JsonObject test)
-    {
-        if (ReadString(test, "testId") is { Length: > 0 } testId)
-        {
-            return $"testId\u001f{testId}";
-        }
-
-        if (ReadString(test, "id") is { Length: > 0 } id)
-        {
-            return $"id\u001f{id}";
-        }
-
-        // `extra` is free-form, so a foreign producer may well have written a string or an array there; indexing
-        // anything but an object by property name throws.
-        if (test["extra"] is JsonObject extra && ReadString(extra, "uid") is { Length: > 0 } uid)
-        {
-            return $"uid\u001f{uid}";
-        }
-
-        if (ReadString(test, "name") is not { Length: > 0 } name)
-        {
-            return null;
-        }
-
-        var identity = new StringBuilder("name");
-        if (test["suite"] is JsonArray suite)
-        {
-            foreach (JsonNode? segment in suite)
-            {
-                // A suite segment is normally a string, but the document is untrusted. Fall back to the
-                // segment's JSON text so a non-string segment still contributes a distinct, deterministic part
-                // of the key instead of throwing on the string conversion.
-                AppendIdentityComponent(
-                    identity,
-                    segment is JsonValue segmentValue && segmentValue.TryGetValue(out string? segmentText)
-                        ? segmentText
-                        : segment?.ToJsonString());
-            }
-        }
-
-        AppendIdentityComponent(identity, name);
-        AppendIdentityComponent(identity, ReadString(test, "filePath"));
-        AppendIdentityComponent(identity, test["parameters"]?.ToJsonString());
-        return identity.ToString();
-    }
-
-    private static void AppendIdentityComponent(StringBuilder identity, string? component)
-        => identity.Append('\u001f').Append(component?.Length ?? -1).Append(':').Append(component);
-
-    private static JsonNode BuildCollapsedTest(JsonObject final, List<JsonObject> priors)
-    {
-        var collapsed = (JsonObject)final.DeepClone();
-
-        // Nothing was merged into this row, so it is passed through rather than synthesized: its own
-        // `retryAttempts[]` (from in-process retries the producer already recorded) stays exactly as written.
-        // Below, once there ARE priors, the history becomes the merger's own array — it has to be rebuilt and
-        // renumbered into a contiguous 1..N-1 — which is why those entries are reshaped and these are not.
-        if (priors.Count == 0)
-        {
-            return collapsed;
-        }
-
-        var history = new JsonArray();
-        foreach (JsonObject prior in priors)
-        {
-            AppendAttempts(history, prior);
-        }
-
-        // In-process retries observed by the final attempt precede its own outcome in the history.
-        if (collapsed["retryAttempts"] is JsonArray finalAttempts)
-        {
-            AppendNestedAttempts(history, finalAttempts);
-        }
-
-        bool anyFailed = false;
-        for (int i = 0; i < history.Count; i++)
-        {
-            if (history[i] is not JsonObject attempt)
-            {
-                continue;
-            }
-
-            attempt["attempt"] = i + 1;
-            anyFailed |= ReadString(attempt, "status") == "failed";
-        }
-
-        collapsed["retryAttempts"] = history;
-        collapsed["retries"] = history.Count;
-
-        // CTRF 9.22: flaky only when the FINAL status is passed after at least one failed attempt. Recomputed
-        // (rather than inherited) because an input's own flag only describes the attempt that produced it.
-        if (ReadString(collapsed, "status") == "passed" && anyFailed)
-        {
-            collapsed["flaky"] = true;
-        }
-        else
-        {
-            collapsed.Remove("flaky");
-        }
-
-        return collapsed;
-    }
-
-    /// <summary>
-    /// Appends the executions a non-final attempt row represents — the attempts it already nested, then its own
-    /// outcome — to the retry history being built.
-    /// </summary>
-    private static void AppendAttempts(JsonArray history, JsonObject test)
-    {
-        if (test["retryAttempts"] is JsonArray nested)
-        {
-            AppendNestedAttempts(history, nested);
-        }
-
-        history.Add(ToRetryAttempt(test));
-    }
-
-    /// <summary>
-    /// Copies the retry attempts an input already recorded into the history being built, projecting each through
-    /// the same section 11 shaping as a promoted test row so a foreign producer's nested attempt cannot smuggle a
-    /// wrong-typed status or a disallowed field into the merged history. Entries that are not objects are skipped:
-    /// they could not describe an execution.
-    /// </summary>
-    private static void AppendNestedAttempts(JsonArray history, JsonArray nested)
-    {
-        foreach (JsonNode? attempt in nested)
-        {
-            if (attempt is JsonObject attemptObject)
-            {
-                history.Add(ToRetryAttempt(attemptObject));
-            }
-        }
-    }
-
-    /// <summary>
-    /// Projects a test row onto the retry attempt object of CTRF section 11, which allows a narrower set of
-    /// fields than a test: everything else (name, suite, tags, labels, ...) either belongs to the collapsed test
-    /// object or, like <c>rawStatus</c>, moves under <c>extra</c>, the only permitted extension point.
-    /// </summary>
-    private static JsonNode ToRetryAttempt(JsonObject test)
-    {
-        // 'attempt' is assigned by the caller, once the position of this execution in the history is known.
-        var attempt = new JsonObject
-        {
-            ["attempt"] = 1,
-            ["status"] = ReadStatus(test),
-        };
-
-        foreach (string field in RetryAttemptFields)
-        {
-            if (test[field] is JsonNode value)
-            {
-                attempt[field] = value.DeepClone();
-            }
-        }
-
-        JsonObject? extra = test["extra"] is JsonObject testExtra ? (JsonObject)testExtra.DeepClone() : null;
-        if (test["rawStatus"] is JsonNode rawStatus)
-        {
-            extra ??= [];
-            extra["rawStatus"] = rawStatus.DeepClone();
-        }
-
-        if (extra is { Count: > 0 })
-        {
-            attempt["extra"] = extra;
-        }
-
-        return attempt;
-    }
-
-    /// <summary>
-    /// Reads a string property, treating a value of any other JSON type as absent. The explicit
-    /// <c>(string?)</c> conversion on a <see cref="JsonNode"/> THROWS for a non-string value rather than
-    /// yielding <see langword="null"/>, and every document reaching the merger is untrusted.
-    /// </summary>
-    private static string? ReadString(JsonObject owner, string propertyName)
-        => owner[propertyName] is JsonValue value && value.TryGetValue(out string? text) ? text : null;
-
-    /// <summary>
-    /// Reads a CTRF status, normalizing anything outside the vocabulary section 11.3 allows — a wrong-typed
-    /// value, or a status this producer invented — to <c>other</c>. Copying such a value verbatim into a retry
-    /// attempt would make the merged document schema-invalid, and <c>other</c> is both a legal status and the
-    /// bucket the summary already counts it in.
-    /// </summary>
-    private static string ReadStatus(JsonObject owner)
-    {
-        string? status = ReadString(owner, "status");
-        return status is "passed" or "failed" or "skipped" or "pending" or "other" ? status : "other";
-    }
-
-    /// <summary>
-    /// Reads an integral property from a JSON node, tolerating anything the node may actually be. The node comes
-    /// straight out of an untrusted document — <c>results.summary</c> is not guaranteed to be an object — so a
-    /// non-object must read as "absent" rather than throw: indexing a <see cref="JsonValue"/> by property name is
-    /// an <see cref="InvalidOperationException"/>.
-    /// </summary>
-    private static bool TryReadLong(JsonNode node, string propertyName, out long value)
-    {
-        value = 0;
-        if (node is not JsonObject jsonObject || jsonObject[propertyName] is not JsonValue jsonValue)
-        {
-            return false;
-        }
-
-        if (jsonValue.TryGetValue(out long longValue))
-        {
-            value = longValue;
-            return true;
-        }
-
-        if (jsonValue.TryGetValue(out double doubleValue))
-        {
-            value = (long)doubleValue;
-            return true;
-        }
-
-        return false;
-    }
-
-    /// <summary>
-    /// Builds a merged environment containing only the fields that every input's environment agrees on
-    /// (so a value that differs across CI agents is dropped rather than attributed to all tests). The
-    /// module-specific <c>extra.testApplication</c> and <c>extra.exitCode</c> fields are always dropped.
-    /// Returns <see langword="null"/> when no environment survives.
-    /// </summary>
-    private static JsonObject? BuildCommonEnvironment(IReadOnlyList<JsonObject> environments, int reportCount)
-    {
-        // A field can only be common to all inputs if every accepted report supplied an environment; a
-        // report with no environment is a disagreement (the field is absent there).
-        if (environments.Count == 0 || environments.Count != reportCount)
-        {
-            return null;
-        }
-
-        var merged = new JsonObject();
-        foreach (KeyValuePair<string, JsonNode?> field in environments[0])
-        {
-            if (field.Key == "extra")
-            {
-                continue;
-            }
-
-            string firstValue = field.Value?.ToJsonString() ?? "null";
-            if (environments.All(e => (e[field.Key]?.ToJsonString() ?? "null") == firstValue))
-            {
-                merged[field.Key] = field.Value?.DeepClone();
-            }
-        }
-
-        if (environments[0]["extra"] is JsonObject firstExtra)
-        {
-            var extra = new JsonObject();
-            foreach (KeyValuePair<string, JsonNode?> field in firstExtra)
-            {
-                if (field.Key is "testApplication" or "exitCode")
-                {
-                    continue;
-                }
-
-                string firstValue = field.Value?.ToJsonString() ?? "null";
-                if (environments.All(e => e["extra"] is JsonObject extraObject && (extraObject[field.Key]?.ToJsonString() ?? "null") == firstValue))
-                {
-                    extra[field.Key] = field.Value?.DeepClone();
-                }
-            }
-
-            if (extra.Count > 0)
-            {
-                merged["extra"] = extra;
-            }
-        }
-
-        return merged.Count > 0 ? merged : null;
-    }
-
-    /// <summary>
-    /// Derives a stable <c>reportId</c> from the accepted CTRF input reports and the merge mode, so identical
-    /// inputs merged the same way reproduce the same id on every retry (RFC 018 idempotency) without a random
-    /// source or reusing an input report's id, while the same inputs merged a DIFFERENT way — which yields a
-    /// materially different document — get a distinct id, as CTRF 5.3 requires. Only the payloads that passed
-    /// CTRF validation are hashed, so a rejected non-CTRF input cannot alter the merged report's identity. A
-    /// non-cryptographic 128-bit FNV-1a fill is sufficient here — the id only needs to be deterministic and
-    /// collision-resistant enough to identify a merged report, not secret.
-    /// </summary>
-    private static string CreateDeterministicReportId(IReadOnlyList<string> acceptedReports, CtrfMergeMode mode)
-    {
-        const ulong fnvPrime = 1099511628211UL;
-        ulong hashLow = 14695981039346656037UL;
-        ulong hashHigh = 0x9E3779B97F4A7C15UL;
-
-        // Fold in the merge mode: the same inputs concatenated and collapsed are two materially different
-        // documents, and CTRF 5.3 wants a distinct reportId for each rather than one id naming both.
-        hashLow = (hashLow ^ (ulong)mode) * fnvPrime;
-        hashHigh = (hashHigh ^ ((ulong)mode + 1UL)) * fnvPrime;
-
-        foreach (string report in acceptedReports)
-        {
-            foreach (char c in report)
-            {
-                hashLow = (hashLow ^ c) * fnvPrime;
-                hashHigh = (hashHigh ^ c) * fnvPrime;
-            }
-
-            // Fold in each input's length so different chunk boundaries (e.g. ["ab","c"] vs ["a","bc"])
-            // never collide.
-            hashLow = (hashLow ^ (ulong)report.Length) * fnvPrime;
-            hashHigh = (hashHigh ^ ((ulong)report.Length + 1UL)) * fnvPrime;
-        }
-
-        byte[] bytes = new byte[16];
-        BitConverter.GetBytes(hashLow).CopyTo(bytes, 0);
-        BitConverter.GetBytes(hashHigh).CopyTo(bytes, 8);
-        return new Guid(bytes).ToString("D");
-    }
-
-    private static long Min(long? current, long candidate)
-        => current is null || candidate < current ? candidate : current.Value;
-
-    private static long Max(long? current, long candidate)
-        => current is null || candidate > current ? candidate : current.Value;
 }
